Changed the endTest message score variable name.
The message displayed to the user at the end of the test previous said '# reached 2:00 average time', time used is an example. This has been changed to reach '# correct 2:00 average time' as that more correctly explains what the number represents. 

This number increments each time a pattern is done correctly rather than hold the highest number of tiles that were on the board. It was called currentMax, perhaps with the intention of counting score a different way. However it has been renamed to correct to better reflect how it is used.