Moved check for chase complete
Previously the check for the user to catch the target was done before the move was checked for validity. This allowed users to touch the target from anywhere, which I don't think is the intended behaviour. The check for target click is now inside the valid move check which means the user can only click on the target if they are beside it.

There is a problem with this change in that it makes the test impossible to complete as the target always moves away from the user. This inability to complete the test needs to be looked at and discussed further.