Added proper catch checking
Previously to catch the target the user just needed to tap on the target, they did not need to be beside the target at all. Now the user needs to be beside the target. The rules of the game have also been changed to allow the user to actually catch the target. Now the user can catch, but not move, diagonally.